Checkout T2: cart CheckoutCarts capability, checked-out retry, Long totals

+++ backend/modules/cart/src/shop/voenix/cart/CartView.kt
@@ -13,15 +13,19 @@ import kotlinx.serialization.Serializable
  * The totals are aggregates of [items] and therefore travel with them: [subtotal] is the sum of
  * article and prompt price times quantity, [shippingCost] and [discountAmount] follow the rules of
  * [CartTotals], and [total] is `subtotal + shippingCost - discountAmount`.
+ *
+ * They are `Long` while a single line price is `Int`, because a line price is one column and a
+ * total is up to 99 of them added up: the sum leaves 32 bits long before any single price does
+ * (deviation D13). The JSON is unchanged — a number is a number.
  */
 @Serializable
 internal data class CartView(
     val id: Long?,
     val items: List<CartLine>,
-    val subtotal: Int,
-    val shippingCost: Int,
-    val discountAmount: Int,
-    val total: Int,
+    val subtotal: Long,
+    val shippingCost: Long,
+    val discountAmount: Long,
+    val total: Long,
     val totalItems: Int,
     val appliedPromotion: AppliedPromotion?,
 ) {
